Remove deleted file
